Tests: Infrastructure: Almost complete refactor
Roughly correct. DB connection doesn't work.
