+++ .travis.yml
@@ -1,40 +1,28 @@
 language: php
 
 php:
-  - 5.4
+  # - 5.5
+  - 5.6
 
 env:
     - WP_VERSION=latest WP_MULTISITE=0
-
-addons:
-  hosts:
-    - local.wordpress.dev
+    # - WP_VERSION=latest WP_MULTISITE=1
 
 before_install:
   # These variables have to be set here, so they can be used by
   # prepare.sh as well as the commands in the script block below
-  - export WORDPRESS_SITE_DIR="$(dirname $TRAVIS_BUILD_DIR)/wordpress/"
-  - export WORDPRESS_TEST_SUBJECT=$(basename $TRAVIS_BUILD_DIR)
-  - export WP_CLI="${TRAVIS_BUILD_DIR}/vendor/bin/wp"
   - export WORDPRESS_FAKE_MAIL_DIR="$(dirname $TRAVIS_BUILD_DIR)/fake-mail/"
-  - echo $WORDPRESS_FAKE_MAIL_DIR
 
 install:
-  - sudo apt-get update > /dev/null
-  - ./ci/prepare.sh
-
-before_script:
+    - bash bin/install-behat.sh
+    - bash bin/install-phpunit.sh wordpress_test root '' localhost $WP_VERSION
 
 script:
-
-  # PHP Lint
-  - make lint
-
-  - make phpunit
-
-  # Run behat tests.
-  - ${WORDPRESS_SITE_DIR}/wp-content/mu-plugins/${WORDPRESS_TEST_SUBJECT}/vendor/bin/behat -c $WORDPRESS_SITE_DIR/wp-content/mu-plugins/${WORDPRESS_TEST_SUBJECT}/behat.yml
+    - if [[ "$WP_VERSION" == "latest" ]]; then if find . -not \( -path ./vendor -prune \) -not \( -path ./features -prune \) -name "*.php" -exec php -l {} \; | grep "^[Parse error|Fatal error]"; then exit 1; fi; fi;
+    - phpunit
+    - bash bin/behat.sh
 
 notifications:
-
-  slack: a8c:rdeQR68xgiffkssb1ZNw5P1c
+    on_success: never
+    on_failure: never
+    # slack: a8c:rdeQR68xgiffkssb1ZNw5P1c
