+++ bin/wp
@@ -1,0 +1,1 @@
+../vendor/wp-cli/wp-cli/bin/wp
