--- ci/wordpress-apache.conf
@@ -1,21 +1,0 @@
-<VirtualHost *:80>
-
-  ServerName local.wordpress.dev
-  DocumentRoot %WORDPRESS_SITE_DIR%
-
-  <Directory "%WORDPRESS_SITE_DIR%">
-    Options FollowSymLinks MultiViews ExecCGI
-    AllowOverride All
-    Order deny,allow
-    Allow from all
-  </Directory>
-
-  # Wire up Apache to use Travis CI's php-fpm.
-  <IfModule mod_fastcgi.c>
-    AddHandler php5-fcgi .php
-    Action php5-fcgi /php5-fcgi
-    Alias /php5-fcgi /usr/lib/cgi-bin/php5-fcgi
-    FastCgiExternalServer /usr/lib/cgi-bin/php5-fcgi -host 127.0.0.1:9000 -pass-header Authorization
-  </IfModule>
-
-</VirtualHost>
