+++ composer.json
@@ -5,12 +5,24 @@
   "require": {
     "php": ">=5.2.4"
   },
+  "require": {
+    "composer/installers": "~1.0"
+  },
   "require-dev": {
-    "behat/behat": "2.4.*@stable",
-    "behat/mink": "1.4.*@stable",
-    "behat/mink-extension": "*",
-    "behat/mink-goutte-driver": "*",
+    "behat/behat": "~3.0",
+    "behat/mink-extension": "~2.0",
+    "behat/mink-goutte-driver": "~1.1",
     "behat/mink-selenium2-driver": "*",
+    "johnbillion/wordpress-behat-extension": "0.1.2",
+    "johnpbloch/wordpress": ">=4.2.0@stable",
+    "phpunit/phpunit": ">=3.7@stable",
     "wp-cli/wp-cli": "*"
+  },
+  "minimum-stability": "dev",
+  "config": {
+    "bin-dir": "bin/"
+  },
+  "extra": {
+    "wordpress-install-dir": "vendor/wordpress"
   }
 }
